+++ test/test-consistency-with-copier-cli.jl
@@ -1,12 +1,12 @@
 @testset "Testing copy, recopy and rebase" begin
   _with_tmp_dir() do dir_copier
-    run(`copier copy --vcs-ref HEAD --quiet $(C.args.copier.ask) $(C.template_path) .`)
+    run(`copier copy --vcs-ref HEAD --quiet $(C.args.copier.robust) $(C.template_path) .`)
     _git_setup()
     _full_precommit()
 
     @testset "Compare copied project vs copier CLI baseline" begin
       _with_tmp_dir() do dir_bestie
-        BestieTemplate.Copier.copy(dir_bestie, C.args.bestie.ask; quiet = true, vcs_ref = "HEAD")
+        BestieTemplate.Copier.copy(dir_bestie, C.args.bestie.robust; quiet = true, vcs_ref = "HEAD")
         _git_setup()
         _full_precommit()
         _test_diff_dir(dir_bestie, dir_copier)
@@ -16,11 +16,11 @@
     @testset "Compare recopied project vs copier CLI baseline" begin
       _with_tmp_dir() do dir_bestie
         run(
-          `copier copy --vcs-ref HEAD --defaults --quiet $(C.args.copier.min) $(C.template_path) .`,
+          `copier copy --vcs-ref HEAD --defaults --quiet $(C.args.copier.minimalistic) $(C.template_path) .`,
         )
         BestieTemplate.Copier.recopy(
           dir_bestie,
-          C.args.bestie.ask;
+          C.args.bestie.robust;
           quiet = true,
           overwrite = true,
           vcs_ref = "HEAD",
@@ -33,11 +33,11 @@
 
     @testset "Compare updated project vs copier CLI baseline" begin
       _with_tmp_dir() do dir_bestie
-        run(`copier copy --defaults --quiet $(C.args.copier.min) $(C.template_path) .`)
+        run(`copier copy --defaults --quiet $(C.args.copier.minimalistic) $(C.template_path) .`)
         _git_setup()
         BestieTemplate.Copier.update(
           dir_bestie,
-          C.args.bestie.ask;
+          C.args.bestie.robust;
           overwrite = true,
           quiet = true,
           vcs_ref = "HEAD",
@@ -50,26 +50,15 @@
   end
 end
 
-@testset "Compare BestieTemplate.generate vs copier CLI on URL/main" begin
-  _with_tmp_dir() do dir_copier
-    run(`copier copy --vcs-ref main --quiet $(C.args.copier.ask) $(C.template_url) .`)
-
-    _with_tmp_dir() do dir_bestie
-      BestieTemplate.generate(dir_bestie, C.args.bestie.ask; quiet = true, vcs_ref = "main")
-      _test_diff_dir(dir_bestie, dir_copier)
-    end
-  end
-end
-
 @testset "Compare BestieTemplate.generate vs copier CLI on HEAD" begin
   _with_tmp_dir() do dir_copier
-    run(`copier copy --vcs-ref HEAD --quiet $(C.args.copier.ask) $(C.template_path) .`)
+    run(`copier copy --vcs-ref HEAD --quiet $(C.args.copier.robust) $(C.template_path) .`)
 
     _with_tmp_dir() do dir_bestie
       BestieTemplate.generate(
         C.template_path,
         dir_bestie,
-        C.args.bestie.ask;
+        C.args.bestie.robust;
         quiet = true,
         vcs_ref = "HEAD",
       )
@@ -91,7 +80,7 @@ end
   _with_tmp_dir() do dir_copier
     _basic_new_pkg("NewPkg")
     run(
-      `copier copy --overwrite --quiet --vcs-ref HEAD $(C.args.copier.min) $(C.template_path) NewPkg`,
+      `copier copy --overwrite --quiet --vcs-ref HEAD $(C.args.copier.minimalistic) $(C.template_path) NewPkg`,
     )
     _fix_project_toml("NewPkg")
 
@@ -100,7 +89,7 @@ end
       BestieTemplate.apply(
         C.template_path,
         joinpath(dir_bestie, "NewPkg"),
-        C.args.bestie.min;
+        C.args.bestie.minimalistic;
         overwrite = true,
         quiet = true,
         vcs_ref = "HEAD",
@@ -113,14 +102,14 @@ end
 
 @testset "Compare BestieTemplate.update vs copier CLI update" begin
   _with_tmp_dir() do dir_copier
-    run(`copier copy --defaults --quiet $(C.args.copier.min) $(C.template_url) .`)
+    run(`copier copy --defaults --quiet $(C.args.copier.minimalistic) $(C.template_url) .`)
     _git_setup()
-    run(`copier update --defaults --quiet $(C.args.copier.ask) .`)
+    run(`copier update --defaults --quiet $(C.args.copier.robust) .`)
 
     _with_tmp_dir() do dir_bestie
-      BestieTemplate.generate(dir_bestie, C.args.bestie.min; defaults = true, quiet = true)
+      BestieTemplate.generate(dir_bestie, C.args.bestie.minimalistic; defaults = true, quiet = true)
       _git_setup()
-      BestieTemplate.update(C.args.bestie.ask; defaults = true, quiet = true)
+      BestieTemplate.update(C.args.bestie.robust; defaults = true, quiet = true)
 
       _test_diff_dir(dir_bestie, dir_copier)
     end